books/requirements.txt
django-csp
django-cors-headers
